demand table modification

+++ estate-services/src/main/resources/db/migration/common/V20201327112043__alter_demand_table.sql
@@ -14,7 +14,7 @@ ALTER TABLE cs_ep_demand
 ADD COLUMN IF NOT EXISTS collected_rent_penalty numeric(13,6);
 
 ALTER TABLE cs_ep_demand 
-ADD COLUMN IF NOT EXISTS interest_since numeric(13,6);
+ADD COLUMN IF NOT EXISTS interest_since bigint;
 
 --Audit table
 
@@ -34,4 +34,4 @@ ALTER TABLE cs_ep_demand_audit
 ADD COLUMN IF NOT EXISTS collected_rent_penalty numeric(13,6);
 
 ALTER TABLE cs_ep_demand_audit 
-ADD COLUMN IF NOT EXISTS interest_since numeric(13,6);
+ADD COLUMN IF NOT EXISTS interest_since bigint;
